+++ test/test_coil.py
@@ -35,21 +35,6 @@ def test_coil_custom_coil():
     # Define a custom coil in testing_data
 
 
-def test_convert_to_mp_unknown_scanner(caplog):
-    dac_units = [14436, 14265, 14045, 9998, 9998, 9998, 9998, 9998]
-
-    convert_to_mp(dac_units, 'unknown')
-    assert "Manufacturer model unknown not implemented, bounds might not be respected. Setting initial " \
-           "shim_setting to 0" in caplog.text
-
-
-def test_convert_to_mp_outside_bounds():
-    dac_units = [20000, 14265, 14045, 9998, 9998, 9998, 9998, 9998]
-
-    with pytest.raises(ValueError, match="Multipole values exceed known system limits."):
-        convert_to_mp(dac_units, 'Prisma_fit')
-
-
 def test_create_scanner_coil_order0():
     sph_contraints = json.load(open(__dir_config_scanner_constraints__))
